+++ main_sketch_GearStickForMargus/motor.h
@@ -10,15 +10,15 @@
 #define MOTOR_SELFCENTER_KD_Q8 (int16_t)(0.1  * SCALE_Q8)  // ~26
 
 void Motor_init(void);
-void Motor_set(int force);
+void Motor_set(int16_t force);
 
 // legacy name kept for compatibility
-void setMotor(int force);
+void setMotor(int16_t force);
 
 // FFB helpers
-int Motor_rampForceToPWM(int rawForce);
-int Motor_computeEndpointPWM(int currentPosition);
-void Motor_selfCenter(int wheelOutput);
+int16_t Motor_rampForceToPWM(int16_t rawForce);
+int16_t Motor_computeEndpointPWM(int32_t currentPosition);
+void Motor_selfCenter(int32_t wheelOutput);
 
 // Diagnostic
 int Motor_getLastForce(void);
